requirements.txt
Django~=3.2.3
mysqlclient~=2.1.1
pymysql~=1.0.3
requests>=2.28.2
djangorestframework-jwt~=1.11.0
cos-python-sdk-v5~=1.9.24
pillow~=10.0.0
djangorestframework~=3.14.0
HTMLParser>=0.0.2
PyQRCode>=1.2.1
wechaty>=0.10.7
requests-toolbelt~=1.0.0
pandas~=1.4.4
openai>=0.27.2
aliyun-python-sdk-core
aliyun-python-sdk-alimt
pydub~=0.25.1
mutagen~=1.47.0
beautifulsoup4~=4.12.2